+++ app/assets/jsx/dashboardPerformer.jsx
@@ -1,98 +1,9 @@
 /** @jsx React.DOM */
 
 define(['react', 'jquery', 'moment'], function(React, $, Moment) {
-  var LoadStatusMixin = {
-    getInitialState: function() {
-      return {
-        buildCI: [],
-      	buildNightly: [],
-        users:[],
-        project:[],
-        audits:[],
-        lastBuild:[],
-        employeesAustria:""
-      };
-    },
-
-    // load data from jenkins, austria and phabricator
-    loadStatus: function() {
-      $.ajax({
-        url: '/buildMain',
-        dataType: 'json',
-        success: function(data1) {
-          this.setState(data1);
-        }.bind(this),
-        error: function(xhr, status, err) {
-          console.error(this.props.url, status, err.toString());
-        }.bind(this)
-      });
-      
-      $.ajax({
-          url: '/buildAside',
-          dataType: 'json',
-          success: function(data1) {
-            this.setState(data1);
-          }.bind(this),
-          error: function(xhr, status, err) {
-            console.error(this.props.url, status, err.toString());
-          }.bind(this)
-        });
-      
-      $.ajax({
-        url: '/getPhabUser',
-        dataType: 'json',
-        success: function(data1) {
-          this.setState(data1);
-        }.bind(this),
-        error: function(xhr, status, err) {
-          console.error(this.props.url, status, err.toString());
-        }.bind(this)
-      });
-      
-      $.ajax({
-        url: '/getPhabProject',
-        dataType: 'json',
-        success: function(data1) {
-          this.setState(data1);
-        }.bind(this),
-        error: function(xhr, status, err) {
-          console.error(this.props.url, status, err.toString());
-        }.bind(this)
-      });
-      
-      $.ajax({
-        url: '/getPhabAudits',
-        dataType: 'json',
-        success: function(data1) {
-          this.setState(data1);
-        }.bind(this),
-        error: function(xhr, status, err) {
-          console.error(this.props.url, status, err.toString());
-        }.bind(this)
-      });
-      
-      $.ajax({
-        url: '/getUsers',
-        dataType: 'json',
-        success: function(data1) {
-          this.setState(data1);
-        }.bind(this),
-        error: function(xhr, status, err) {
-          console.error(this.props.url, status, err.toString());
-        }.bind(this)
-      });
-    },
-    
-    componentWillMount: function() {
-      this.loadStatus();
-      setInterval(this.loadStatus, this.props.pollInterval);
-    },
-  };
-
   
   var Dashboard = React.createClass({
-    mixins: [LoadStatusMixin],
-
+    mixins: [getStatusMixin()],
     getDefaultProps: function() {
       return {
         pollInterval: 10000
@@ -100,76 +11,7 @@ define(['react', 'jquery', 'moment'], function(React, $, Moment) {
     },
 
     render: function() {
-      // ----------------------- helper -----------------------//
-      function getAuditsForUser(user, audits){
-        var res = 0;
-        if (!isUserInProject(user)) return res;
-        for (var i = 0; i < audits.length; i++){
-          if (user.phid == audits[i].auditorPHID &&
-              audits[i].status != "accepted"){
-            res++;
-          }
-        }
-        return res;
-      }
-      var project = this.state.project;
-      function isUserInProject(user){
-        var key = Object.keys(project.data)
-        for (var i = 0; i < project.data[key].members.length; i++)
-          if (project.data[key].members[i] == user.phid)
-            return true;
-        return false;
-      }
-      function getNotAssignedAudits(audits){
-        var res = 0;
-        for(var i = 0; i<audits.length; i++){
-            if (audits[i].status === "audit-required" &&
-                audits[i].auditorPHID == Object.keys(project.data)){
-              res++;
-          }
-        }
-        return res;
-      }
-      
-      function mergeUserAudits(users, audits){
-        var resultArray = new Array();
-        for (var i = 0; i < users.length; i++){
-          var cnt = getAuditsForUser(users[i], audits);
-          if (cnt > 0){
-            var userAudit = {
-                userName : users[i].realName,
-                numberOfAudits : cnt
-            };
-            resultArray.push(userAudit);
-          }
-        }
-        
-        var userAudit = {
-            userName : "No Auditor",
-            numberOfAudits : getNotAssignedAudits(audits)
-        };
-        resultArray.push(userAudit);
-        
-        return resultArray;
-      }
-      
-      function getDefaultPicture(){
-        return '/assets/images/avatars/default.jpg';
-      }
-      
-      function formatEmplName(name){
-        return name.replace(" ", ".")
-                   .replace(/ä/g,"ae")
-                   .replace(/ö/g,"oe")
-                   .replace(/ü/g,"ue")
-                   .replace(/Ä/g,"Ae")
-                   .replace(/Ö/g,"Oe")
-                   .replace(/Ü/g,"Ue")
-                   .replace(/ß/g,"ss") + '.jpg';
-      }
-      
       var empl = this.state.employeesAustria.toLowerCase();
-      
       function getPicture(name){
         name = formatEmplName(name);
         
@@ -181,8 +23,8 @@ define(['react', 'jquery', 'moment'], function(React, $, Moment) {
         }
         else{
             return 'http://austria/global/images/employees/' +  name;      
-          }
         }
+      }
       
       var phabUsers = this.state.users;
       function getCommitters(committer){
@@ -353,7 +195,6 @@ define(['react', 'jquery', 'moment'], function(React, $, Moment) {
             </li>
         );
       };
-    	
      
       function renderAudit(audit) {
         return (
@@ -371,8 +212,10 @@ define(['react', 'jquery', 'moment'], function(React, $, Moment) {
    // ----------------------- generate html -----------------------//
       var buildItems = this.state.buildCI.map(buildItems);
       var buildNightly = this.state.buildNightly.map(buildItemsNightly);
-      var audits = mergeUserAudits(this.state.users, this.state.audits).sort(function(a, b){return b.numberOfAudits-a.numberOfAudits}).map(renderAudit);
-      
+      var audits = mergeUserAudits(this.state.users, this.state.audits, this.state.project).sort(function(a, b){return b.numberOfAudits-a.numberOfAudits}).map(renderAudit);
+      var iFrameStyle = {
+          height: '100%'
+        }
    // ----------------------- html site structure -----------------------//
       return (
         <div>
@@ -382,7 +225,6 @@ define(['react', 'jquery', 'moment'], function(React, $, Moment) {
                 {buildItems}
             	</ul>
             </section>
-            
             <section className="audit-section">
               <h1> Open Audits </h1>
               <ul className="audit-list">
@@ -391,13 +233,10 @@ define(['react', 'jquery', 'moment'], function(React, $, Moment) {
             </section>
             <aside id="nightly-build" className="nightly-build">
               {buildNightly} 
-              <iframe className="failureSummary" src="http://lnz-spbuilder/jenkins/job/Status_all_Core_Regressions/Core_Regressions_Status/output.html" scrolling="no" 
-                seamless="seamless"/>
+              <iframe style={iFrameStyle} className="failureSummary" src="http://lnz-spbuilder/jenkins/job/Status_all_Core_Regressions/Core_Regressions_Status/output.html" scrolling="no"/>
             </aside>
-            
           </article> 
         </div>
-        
       );
     }
   });
