+++ app/util/JenkinsFetcherPerformer.scala
@@ -6,11 +6,6 @@ import play.api.libs.ws.WS
 import play.api.Play.current
 import play.api.libs.json._
 import scala.concurrent.ExecutionContext.Implicits.global
-import java.util.Date
-import play.Logger
-import models.NevergreenResult
-import scala.util.parsing.json.JSONArray
-import scala.util.Try
 
 object JenkinsFetcherPerformer extends JenkinsFetcher {
