+++ app/util/JenkinsFetcherUtil.scala
@@ -1,16 +1,11 @@
 package util
 
 import scala.concurrent.Future
-import play.api.Play
 import play.api.libs.ws.WS
 import play.api.Play.current
 import play.api.libs.json._
 import scala.concurrent.ExecutionContext.Implicits.global
-import java.util.Date
-import play.Logger
-import models.NevergreenResult
-import scala.util.parsing.json.JSONArray
-import scala.util.Try
+
 object JenkinsFetcherUtil {
 
   def getDetails(json: JsValue, numberOfItems: Integer, baseUrl: String): Future[List[JsObject]] = {
