+++ app/views/main.scala.html
@@ -8,6 +8,8 @@
         <link rel="stylesheet" media="screen" href="@routes.Assets.versioned("stylesheets/dashboard.css")">
         <link rel="shortcut icon" type="image/png" href="@routes.Assets.versioned("images/favicon.png")">
         <script data-main="@routes.Assets.versioned(jsFile)" type="text/javascript" src="@routes.Assets.versioned("lib/requirejs/require.js")"></script>
+        <script data-main="@routes.Assets.versioned(jsFile)" type="text/javascript" src="@routes.Assets.versioned("jsx/dashboardLoader.js")"></script>
+        <script data-main="@routes.Assets.versioned(jsFile)" type="text/javascript" src="@routes.Assets.versioned("jsx/dashboardHelper.js")"></script>
     </head>
     <body>
         @content
