+++ conf/application.conf
@@ -65,8 +65,8 @@ logger.application=DEBUG
 #dashboard.mockResponse=true
 # Phabricator config
 
-#dashboard.fetcher="performer"
-dashboard.fetcher="silktest"
+dashboard.fetcher="performer"
+#dashboard.fetcher="silktest"
 
 dashboard.urlPhabricator="http://lnz-phabricator.microfocus.com/"
